+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 ## Changelog ##
 
 
+### 0.13.2 ###
+
+Update database after plugin update
+
+* Now you don't have to deactivate and activate plugin after every update!
+* Tested on standalone mock-plugin
+
+
 ### 0.13.1 ###
 
 Fix result permalinks with front prefixes
@@ -50,6 +58,7 @@ Speed-up tests saving, improve respondents results' and add new locale
 
 
 ### 0.11.4 ###
+
 Fix content comparing in duplicate protection
 
 You should upgrade to this version only if:
@@ -60,6 +69,7 @@ You should upgrade to this version only if:
 
 
 ### 0.11.3 ###
+
 Fix strict settings catchable error in steps
 
 Upgrade notice:
@@ -68,12 +78,14 @@ Upgrade notice:
 
 
 ### 0.11.2 ###
+
 Fix many answers on many steps
 
 * Was stuck on 2nd step
 
 
 ### 0.11.1 ###
+
 Improve steps generating (for sections addon)
 
 Upgrade notice:
@@ -82,6 +94,7 @@ Upgrade notice:
 
 
 ### 0.11 ###
+
 Add one-per-page questions and fix annoing database collation issue
 
 * Add option when respondent could read only one question at a time on page, to avoid to see through pattern
@@ -186,6 +199,7 @@ Fix external library and update translations
 
 
 ### 0.8 ###
+
 Add test page settings
 
 * Allow to reset answers and customize button caption
@@ -200,10 +214,12 @@ Internal improvements:
 
 
 ### 0.7.1 ###
+
 Fix bad external library version
 
 
 ### 0.7 ###
+
 Give each passing own URL to allow sharing and add two result page options
 
 * Save passings in DB and redirect to them by URLs (with client's ip and device uuid)
@@ -214,22 +230,27 @@ Give each passing own URL to allow sharing and add two result page options
 * Update translations
 
 ### 0.6.4 ###
+
 Italian translation added
 
 
 ### 0.6.3 ###
+
 Fix featured image incompatibilities with Jetpack shortcode module
 
 
 ### 0.6.2 ###
+
 Inherit post's CSS classes onto test
 
 
 ### 0.6.1 ###
+
 Fix apostrophe problem (slashes) and update tests to be compatible to WP 4.1
 
 
 ### 0.6 ###
+
 Individual answers
 
 * Add individual answers feature
@@ -240,22 +261,26 @@ Individual answers
 
 
 ### 0.5.4 ###
+
 Update German translation
 
 
 ### 0.5.3 ###
+
 Add standard category to tests
 
 * Update README about Bulgarian translation
 
 
 ### 0.5.2 ###
+
 Fix fatal error incompatibility with wordpress-seo plugin
 
 * Adding Bulgarian translation
 
 
 ### 0.5.1 ###
+
 Add tags, improve formulas editor and docs
 
 * Prefix screenhots by zeros
@@ -266,6 +291,7 @@ Add tags, improve formulas editor and docs
 
 
 ### 0.5 ###
+
 Localization and quick fill
 
 * Localize to six languages with help of [Transifiex](https://www.transifex.com/projects/p/wp-testing/)
@@ -276,6 +302,7 @@ Localization and quick fill
 
 
 ### 0.4.2 ###
+
 Fix pages disapperance and form formatting
 
 * Fix pages disapperance
@@ -283,12 +310,14 @@ Fix pages disapperance and form formatting
 
 
 ### 0.4.1 ###
+
 Fix tests preview mode
 
 * Not changing main WP query when in preview
 
 
 ### 0.4 ###
+
 Display tests in blog everywhere same as posts
 
 * Display tests on homepage, in categories and other places just like posts
@@ -298,6 +327,7 @@ Display tests in blog everywhere same as posts
 * Minimize possibility of "Max post vars" warning by minizing the number of hidden inputs
 
 ### 0.3 ###
+
 Test plugin functionality in 37 combinations of WP and PHP and fix found problems
 
 * Test under WordPress from 3.2 to latest 4.0 and PHP from 5.2 to 5.5
@@ -325,26 +355,32 @@ Non-latest WP versions fixes (programmer's language):
 
 
 ### 0.2.5 ###
+
 Fix test creation (empty scale sum broken)
 
 
 ### 0.2.4 ###
+
 Correctly uninstall plugin
 
 
 ### 0.2.3 ###
+
 Fix activation on PHP below 5.4
 
 
 ### 0.2.2 ###
+
 Update description to correct English version
 
 
 ### 0.2.1 ###
+
 Fix PHP 5.2 parse error (not affects latest PHP versions)
 
 
 ### 0.2 ###
+
 Connecting scales scores with results through formulas.
 
 * Formulas parser undertands variables and comparision operators like "less", "more", "same", "not same", "and", "or"
@@ -355,6 +391,7 @@ Connecting scales scores with results through formulas.
 
 
 ### 0.1.4 ###
+
 Test passing error fixed
 
 * Manual relashionships naming
@@ -362,6 +399,7 @@ Test passing error fixed
 
 
 ### 0.1.3 ###
+
 Another plugin activation problems fixed
 
 * Permissions on migrations directory
@@ -369,10 +407,12 @@ Another plugin activation problems fixed
 
 
 ### 0.1.2 ###
+
 Bump stable tag to apply previous hotfix on wordpress plugins
 
 
 ### 0.1.1 ###
+
 Plugin activation hotfix
 
 * In initial release migrations dir taken from wordpress dir rather than be hardcoded.
@@ -381,6 +421,7 @@ Locally all was ok as always, but not on your wordpresses :(
 
 
 ### 0.1 ###
+
 Initial release
 
 * Add shortcode for tests lists: wptlist
@@ -397,6 +438,9 @@ Initial release
 
 
 == Upgrade Notice ==
+
+### 0.13.2 ###
+Now you don't have to deactivate and activate plugin after every update
 
 ### 0.7.1 ###
 Fix bad external library version